Move ReadHeader to `utils` file so symbols do not clash
